+++ internal/consumer/clusterConsumer.go
@@ -25,6 +25,7 @@ import (
 	"fmt"
 	"strconv"
 	"sync"
+	"time"
 
 	"github.com/bsm/sarama-cluster"
 	"github.com/uber-go/kafka-client/internal/metrics"
@@ -33,6 +34,10 @@ import (
 	"github.com/uber-go/tally"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
+)
+
+const (
+	metricsInterval = time.Minute
 )
 
 type (
@@ -44,6 +49,7 @@ type (
 		scope            tally.Scope
 		logger           *zap.Logger
 		lifecycle        *util.RunLifecycle
+		metricsTicker    *time.Ticker
 		stopC            chan struct{}
 		doneC            chan struct{}
 	}
@@ -64,6 +70,7 @@ func NewClusterConsumer(
 		scope:            scope.Tagged(map[string]string{"cluster": cluster}),
 		logger:           logger.With(zap.String("cluster", cluster)),
 		lifecycle:        util.NewRunLifecycle(cluster + "-consumer"),
+		metricsTicker:    time.NewTicker(metricsInterval),
 		stopC:            make(chan struct{}),
 		doneC:            make(chan struct{}),
 	}
@@ -114,14 +121,30 @@ func (c *ClusterConsumer) Closed() <-chan struct{} {
 
 // eventLoop is the main event loop for this consumer
 func (c *ClusterConsumer) eventLoop() {
+	var n *cluster.Notification
+	var ok bool
 	for {
 		select {
-		case pc := <-c.consumer.Partitions():
-			c.addPartitionConsumer(pc)
-		case n := <-c.consumer.Notifications():
-			c.handleNotification(n)
-		case err := <-c.consumer.Errors():
-			c.logger.Warn("cluster consumer error", zap.Error(err))
+		case pc, ok := <-c.consumer.Partitions():
+			if ok {
+				c.addPartitionConsumer(pc)
+			}
+		case n, ok = <-c.consumer.Notifications():
+			if ok {
+				c.handleNotification(n)
+			}
+		case err, ok := <-c.consumer.Errors():
+			if ok {
+				c.logger.Warn("cluster consumer error", zap.Error(err))
+			}
+		case _, ok := <-c.metricsTicker.C:
+			if ok && n != nil {
+				for topic, partitions := range n.Current {
+					for _, partition := range partitions {
+						c.scope.Tagged(map[string]string{"topic": topic, "partition": strconv.Itoa(int(partition))}).Gauge(metrics.KafkaPartitionOwned).Update(1.0)
+					}
+				}
+			}
 		case <-c.stopC:
 			c.shutdown()
 			c.logger.Info("cluster consumer stopped")
@@ -182,6 +205,7 @@ func (c *ClusterConsumer) shutdown() {
 	}
 	wg.Wait()
 	c.consumer.Close() // close sarama cluster consumer
+	c.metricsTicker.Stop()
 	close(c.doneC)
 }