+++ internal/consumer/options.go
@@ -30,9 +30,10 @@ import (
 type (
 	// Options are the tunable and injectable options for the consumer
 	Options struct {
-		RcvBufferSize          int // aggregate message buffer size
-		PartitionRcvBufferSize int // message buffer size for each partition
-		Concurrency            int // number of goroutines that will concurrently process messages
+		ClientID               string // client ID
+		RcvBufferSize          int    // aggregate message buffer size
+		PartitionRcvBufferSize int    // message buffer size for each partition
+		Concurrency            int    // number of goroutines that will concurrently process messages
 		OffsetPolicy           int64
 		OffsetCommitInterval   time.Duration
 		RebalanceDwellTime     time.Duration