+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 v0.2.2 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Update KafkaPartitionOwned metric to be a boolean metric for ownership.
+- Rename number of partitions owned by specific worker to KafkaPartitionOwnedCount.
 
 
 v0.2.1 (2018-08-07)