+++ CHANGELOG.md
@@ -4,6 +4,8 @@ v0.2.2 (unreleased)
 -------------------
 
 - Handle retryQ and DLQ only consumer topics.
+- Update KafkaPartitionOwned metric to be a boolean metric for ownership.
+- Rename number of partitions owned by specific worker to KafkaPartitionOwnedCount.
 
 
 v0.2.1 (2018-08-07)